// restaurant-front/src/app/models/Utilisateur.model.ts
export interface Utilisateur {
    id: number; // Maps to the 'id' field in the Java entity
    name: string; // Maps to the 'nom' field in the Java entity
    email: string; // Maps to the 'email' field in the Java entity
    password: string; // Maps to the 'password' field in the Java entity
    role: string; // Maps to the 'Role' field in the Java entity
  }